makes AbstractReadParametersBeforeSet apply to only machine verts
