+++ spynnaker/pyNN/models/abstract_models/abstract_read_parameters_before_set.py
@@ -14,6 +14,9 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from six import add_metaclass
+
+from pacman.model.graphs.machine import MachineVertex
+from spinn_front_end_common.utilities.exceptions import ConfigurationException
 from spinn_utilities.abstract_base import AbstractBase, abstractmethod
 
 
@@ -23,6 +26,17 @@ class AbstractReadParametersBeforeSet(object):
     """
 
     __slots__ = ()
+
+    _WRONG_VERTEX_TYPE_ERROR = (
+        "The vertex {} is not of type MachineVertex. By not being a "
+        "machine vertex, the sPyNNaker population set function may not "
+        "work correctly.")
+
+    def __new__(cls, *args, **kwargs):
+        if not issubclass(cls, MachineVertex):
+            raise ConfigurationException(
+                cls._WRONG_VERTEX_TYPE_ERROR.format(cls))
+        return super(AbstractReadParametersBeforeSet, cls).__new__(cls)
 
     @abstractmethod
     def read_parameters_from_machine(
