+++ spynnaker/pyNN/models/neuron/abstract_population_vertex.py
@@ -37,8 +37,7 @@ from spynnaker.pyNN.models.common import (
 from spynnaker.pyNN.utilities import bit_field_utilities
 from spynnaker.pyNN.models.abstract_models import (
     AbstractPopulationInitializable, AbstractAcceptsIncomingSynapses,
-    AbstractPopulationSettable, AbstractReadParametersBeforeSet,
-    AbstractContainsUnits)
+    AbstractPopulationSettable, AbstractContainsUnits)
 from spynnaker.pyNN.exceptions import InvalidParameterType
 from spynnaker.pyNN.utilities.ranged import (
     SpynnakerRangeDictionary, SpynnakerRangedList)
@@ -58,9 +57,8 @@ class AbstractPopulationVertex(
         AbstractSpikeRecordable, AbstractNeuronRecordable,
         AbstractProvidesOutgoingPartitionConstraints,
         AbstractPopulationInitializable, AbstractPopulationSettable,
-        AbstractChangableAfterRun, AbstractReadParametersBeforeSet,
-        AbstractAcceptsIncomingSynapses, ProvidesKeyToAtomMappingImpl,
-        AbstractCanReset):
+        AbstractChangableAfterRun, AbstractAcceptsIncomingSynapses,
+        ProvidesKeyToAtomMappingImpl, AbstractCanReset):
     """ Underlying vertex model for Neural Populations.
         Not actually abstract.
     """
@@ -513,37 +511,6 @@ class AbstractPopulationVertex(
         for vertex in self.machine_vertices:
             if isinstance(vertex, AbstractRewritesDataSpecification):
                 vertex.set_reload_required(True)
-
-    @overrides(AbstractReadParametersBeforeSet.read_parameters_from_machine)
-    def read_parameters_from_machine(
-            self, transceiver, placement, vertex_slice):
-
-        # locate SDRAM address to where the neuron parameters are stored
-        neuron_region_sdram_address = (
-            placement.vertex.neuron_region_sdram_address(
-                placement, transceiver))
-
-        # shift past the extra stuff before neuron parameters that we don't
-        # need to read
-        neuron_parameters_sdram_address = (
-            neuron_region_sdram_address + self.tdma_sdram_size_in_bytes +
-            self._BYTES_TILL_START_OF_GLOBAL_PARAMETERS)
-
-        # get size of neuron params
-        size_of_region = self.get_sdram_usage_for_neuron_params(vertex_slice)
-        size_of_region -= (
-            self._BYTES_TILL_START_OF_GLOBAL_PARAMETERS +
-            self.tdma_sdram_size_in_bytes)
-
-        # get data from the machine
-        byte_array = transceiver.read_memory(
-            placement.x, placement.y, neuron_parameters_sdram_address,
-            size_of_region)
-
-        # update python neuron parameters with the data
-        self.__neuron_impl.read_data(
-            byte_array, 0, vertex_slice, self._parameters,
-            self._state_variables)
 
     @property
     def weight_scale(self):
