+++ spynnaker/pyNN/models/neuron/population_machine_vertex.py
@@ -38,7 +38,8 @@ from spinn_front_end_common.utilities.utility_objs import ExecutableType
 from spynnaker.pyNN.models.neuron.synapse_dynamics import (
     AbstractSynapseDynamicsStructural)
 from spynnaker.pyNN.utilities import constants, bit_field_utilities
-from spynnaker.pyNN.models.abstract_models import AbstractSynapseExpandable
+from spynnaker.pyNN.models.abstract_models import (
+    AbstractSynapseExpandable, AbstractReadParametersBeforeSet)
 from spynnaker.pyNN.utilities.constants import POPULATION_BASED_REGIONS
 from spinn_front_end_common.utilities import (
     constants as common_constants, helpful_functions)
@@ -51,7 +52,7 @@ class PopulationMachineVertex(
         AbstractSupportsBitFieldGeneration,
         AbstractSupportsBitFieldRoutingCompression,
         AbstractGeneratesDataSpecification, AbstractSynapseExpandable,
-        AbstractRewritesDataSpecification):
+        AbstractRewritesDataSpecification, AbstractReadParametersBeforeSet):
 
     __slots__ = [
         "__binary_file_name",
@@ -627,3 +628,33 @@ class PopulationMachineVertex(
     def read_generated_connection_holders(self, transceiver, placement):
         self._app_vertex.synapse_manager.read_generated_connection_holders(
             transceiver, placement)
+
+    @overrides(AbstractReadParametersBeforeSet.read_parameters_from_machine)
+    def read_parameters_from_machine(
+            self, transceiver, placement, vertex_slice):
+
+        # locate SDRAM address to where the neuron parameters are stored
+        neuron_region_sdram_address = self.neuron_region_sdram_address(
+            placement, transceiver)
+
+        # shift past the extra stuff before neuron parameters that we don't
+        # need to read
+        neuron_parameters_sdram_address = (
+            neuron_region_sdram_address + self.tdma_sdram_size_in_bytes +
+            self._BYTES_TILL_START_OF_GLOBAL_PARAMETERS)
+
+        # get size of neuron params
+        size_of_region = self.get_sdram_usage_for_neuron_params(vertex_slice)
+        size_of_region -= (
+            self._BYTES_TILL_START_OF_GLOBAL_PARAMETERS +
+            self.tdma_sdram_size_in_bytes)
+
+        # get data from the machine
+        byte_array = transceiver.read_memory(
+            placement.x, placement.y, neuron_parameters_sdram_address,
+            size_of_region)
+
+        # update python neuron parameters with the data
+        self.__neuron_impl.read_data(
+            byte_array, 0, vertex_slice, self._parameters,
+            self._state_variables)
