+++ spynnaker/pyNN/models/pynn_population_common.py
@@ -61,8 +61,7 @@ class PyNNPopulationCommon(object):
         "_vertex_changeable_after_run",
         "_vertex_contains_units",
         "_vertex_population_initializable",
-        "_vertex_population_settable",
-        "_vertex_read_parameters_before_set"]
+        "_vertex_population_settable"]
 
     def __init__(
             self, spinnaker_control, size, label, constraints, model,
@@ -137,8 +136,6 @@ class PyNNPopulationCommon(object):
             isinstance(self.__vertex, AbstractPopulationInitializable)
         self._vertex_changeable_after_run = \
             isinstance(self.__vertex, AbstractChangableAfterRun)
-        self._vertex_read_parameters_before_set = \
-            isinstance(self.__vertex, AbstractReadParametersBeforeSet)
         self._vertex_contains_units = \
             isinstance(self.__vertex, AbstractContainsUnits)
 
@@ -507,21 +504,21 @@ class PyNNPopulationCommon(object):
 
         # If the tools have run before, and not reset, and the read
         # hasn't already been done, read back the data
-        if globals_variables.get_simulator().has_ran \
-                and self._vertex_read_parameters_before_set \
-                and not self.__has_read_neuron_parameters_this_run \
-                and not globals_variables.get_simulator().use_virtual_board:
+        if (globals_variables.get_simulator().has_ran
+                and not self.__has_read_neuron_parameters_this_run
+                and not globals_variables.get_simulator().use_virtual_board):
             # go through each machine vertex and read the neuron parameters
             # it contains
             for machine_vertex in self.__vertex.machine_vertices:
-                # tell the core to rewrite neuron params back to the
-                # SDRAM space.
-                placement = globals_variables.get_simulator().placements.\
-                    get_placement_of_vertex(machine_vertex)
+                if isinstance(machine_vertex, AbstractReadParametersBeforeSet):
+                    # tell the core to rewrite neuron params back to the
+                    # SDRAM space.
+                    placement = globals_variables.get_simulator().\
+                        placements.get_placement_of_vertex(machine_vertex)
 
-                self.__vertex.read_parameters_from_machine(
-                    globals_variables.get_simulator().transceiver, placement,
-                    machine_vertex.vertex_slice)
+                    self.__vertex.read_parameters_from_machine(
+                        globals_variables.get_simulator().transceiver,
+                        placement, machine_vertex.vertex_slice)
 
             self.__has_read_neuron_parameters_this_run = True
