+++ spynnaker/pyNN/models/spike_source/spike_source_poisson_machine_vertex.py
@@ -12,19 +12,20 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
+import struct
 from enum import Enum
 
 import numpy
 
 from data_specification.enums import DataType
-from spinn_front_end_common.interface.buffer_management import \
-    recording_utilities
+from spinn_front_end_common.interface.buffer_management import (
+    recording_utilities)
 from spinn_front_end_common.interface.simulation import simulation_utilities
-from spinn_front_end_common.utilities import helpful_functions
-from spinn_front_end_common.utilities.constants import \
-    MICRO_TO_SECOND_CONVERSION, SIMULATION_N_BYTES, BYTES_PER_WORD, \
-    MICRO_TO_MILLISECOND_CONVERSION
+from spinn_front_end_common.utilities import (
+    helpful_functions, globals_variables)
+from spinn_front_end_common.utilities.constants import (
+    MICRO_TO_SECOND_CONVERSION, SIMULATION_N_BYTES, BYTES_PER_WORD,
+    MICRO_TO_MILLISECOND_CONVERSION)
 from spinn_utilities.overrides import overrides
 from pacman.executor.injection_decorator import inject_items
 from pacman.model.graphs.machine import MachineVertex
@@ -44,10 +45,12 @@ from spinn_front_end_common.interface.profiling import (
     AbstractHasProfileData, profile_utils)
 from spinn_front_end_common.interface.profiling.profile_utils import (
     get_profiling_data)
-from spynnaker.pyNN.models.abstract_models import AbstractMaxSpikes
+from spynnaker.pyNN.models.abstract_models import (
+    AbstractMaxSpikes, AbstractReadParametersBeforeSet)
 from spynnaker.pyNN.utilities import constants
 from spynnaker.pyNN.utilities.constants import (
     LIVE_POISSON_CONTROL_PARTITION_ID)
+from spynnaker.pyNN.utilities.struct import Struct
 
 
 def _flatten(alist):
@@ -84,7 +87,7 @@ class SpikeSourcePoissonMachineVertex(
         ProvidesProvenanceDataFromMachineImpl, AbstractRecordable,
         AbstractSupportsDatabaseInjection, AbstractHasProfileData,
         AbstractHasAssociatedBinary, AbstractRewritesDataSpecification,
-        AbstractGeneratesDataSpecification):
+        AbstractGeneratesDataSpecification, AbstractReadParametersBeforeSet):
 
     __slots__ = [
         "__buffered_sdram_per_timestep",
@@ -105,6 +108,16 @@ class SpikeSourcePoissonMachineVertex(
     PROFILE_TAG_LABELS = {
         0: "TIMER",
         1: "PROB_FUNC"}
+
+    _PoissonStruct = Struct([
+        DataType.UINT32,  # Start Scaled
+        DataType.UINT32,  # End Scaled
+        DataType.UINT32,  # Next Scaled
+        DataType.UINT32,  # is_fast_source
+        DataType.U032,  # exp^(-spikes_per_tick)
+        DataType.S1615,  # sqrt(spikes_per_tick)
+        DataType.UINT32,  # inter-spike-interval
+        DataType.UINT32])  # timesteps to next spike
 
     class EXTRA_PROVENANCE_DATA_ENTRIES(Enum):
         """ Entries for the provenance data generated by standard neuron \
@@ -129,6 +142,10 @@ class SpikeSourcePoissonMachineVertex(
     # 8. uint32_t first_source_id; 9. uint32_t n_spike_sources;
     # 10,11,12,13 mars_kiss64_seed_t (uint[4]) spike_source_seed;
     PARAMS_BASE_WORDS = 13
+
+    # Seed offset in parameters and size on bytes
+    SEED_OFFSET_BYTES = 9 * 4
+    SEED_SIZE_BYTES = 4 * 4
 
     def __init__(
             self, resources_required, is_recording, constraints=None,
@@ -614,3 +631,70 @@ class SpikeSourcePoissonMachineVertex(
             placement,
             self.POISSON_SPIKE_SOURCE_REGIONS.RATES_REGION.value,
             transceiver)
+
+    @overrides(
+        AbstractReadParametersBeforeSet.read_parameters_from_machine)
+    def read_parameters_from_machine(
+            self, transceiver, placement, vertex_slice):
+
+        # locate SDRAM address where parameters are stored
+        poisson_params = self.poisson_param_region_address(
+            placement, transceiver)
+        seed_array = transceiver.read_memory(
+            placement.x, placement.y, poisson_params + self.SEED_OFFSET_BYTES,
+            self.SEED_SIZE_BYTES)
+        self.__kiss_seed[vertex_slice] = struct.unpack_from("<4I", seed_array)
+
+        # locate SDRAM address where the rates are stored
+        poisson_rate_region_sdram_address = (
+            self.poisson_rate_region_address(placement, transceiver))
+
+        # get size of poisson params
+        size_of_region = get_rates_bytes(vertex_slice, self.__data["rates"])
+
+        # get data from the machine
+        byte_array = transceiver.read_memory(
+            placement.x, placement.y,
+            poisson_rate_region_sdram_address, size_of_region)
+
+        # For each atom, read the number of rates and the rate parameters
+        offset = 0
+        for i in range(vertex_slice.lo_atom, vertex_slice.hi_atom + 1):
+            n_values = struct.unpack_from("<I", byte_array, offset)[0]
+            offset += 4
+
+            # Skip reading the index, as it will be recalculated on data write
+            offset += 4
+
+            (_start, _end, _next, is_fast_source, exp_minus_lambda,
+             sqrt_lambda, isi, time_to_next_spike) = (
+                 self._PoissonStruct.read_data(byte_array, offset, n_values))
+            offset += (
+                self._PoissonStruct.get_size_in_whole_words(
+                    n_values) * BYTES_PER_WORD)
+
+            # Work out the spikes per tick depending on if the source is
+            # slow (isi), fast (exp) or faster (sqrt)
+            is_fast_source = is_fast_source == 1.0
+            spikes_per_tick = numpy.zeros(len(is_fast_source), dtype="float")
+            spikes_per_tick[is_fast_source] = numpy.log(
+                exp_minus_lambda[is_fast_source]) * -1.0
+            is_faster_source = sqrt_lambda > 0
+            # pylint: disable=assignment-from-no-return
+            spikes_per_tick[is_faster_source] = numpy.square(
+                sqrt_lambda[is_faster_source])
+            slow_elements = isi > 0
+            spikes_per_tick[slow_elements] = 1.0 / isi[slow_elements]
+
+            # Convert spikes per tick to rates
+            machine_time_step = (
+                globals_variables.get_simulator().machine_time_step)
+            self.__data["rates"].set_value_by_id(
+                i,
+                spikes_per_tick *
+                (MICRO_TO_SECOND_CONVERSION / float(machine_time_step)))
+
+            # Store the updated time until next spike so that it can be
+            # rewritten when the parameters are loaded
+            self.__data["time_to_spike"].set_value_by_id(
+                i, time_to_next_spike)
